More details

+++ financetoolkit/base/ratios/ratios_controller.py
@@ -317,9 +317,11 @@ class Ratios:
                         break
 
             if formula_adjusted:
-                calculation = eval(formula_adjusted).astype(np.float64)  # ruff: noqa
+                calculation = eval(formula_adjusted)  # ruff: noqa
 
-                total_financials.loc[:, name, :] = calculation.to_numpy()
+                total_financials.loc[:, name, :] = calculation.astype(
+                    np.float64
+                ).to_numpy()
 
                 self._custom_ratios = total_financials.loc[
                     :, list(custom_ratios_dict.keys()), :
